requirements.txt
fastapi==0.110.0
motor==3.3.2
pymongo==4.6.1
qdrant-client==1.7.0
langchain-google-genai==0.1.7
langchain==0.1.14
langchain-community==0.0.28
langchain-openai==0.0.6
pydantic==2.6.4
python-dotenv==1.0.1
uvicorn==0.29.0
python-multipart==0.0.9